dqs-serve/pyproject.toml
[project]
name = "dqs-serve"
version = "0.1.0"
description = "Data Quality Service — FastAPI serve layer, MCP tools, and schema DDL owner"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "fastmcp>=3.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require a running Postgres instance",
]
addopts = "-m 'not integration'"
